Implemented FormSearchView and ApplicationView

+++ web/src/main/kotlin/sashastory/dev/dao/Sql2oApplicationDao.kt
@@ -4,6 +4,7 @@ import com.github.vok.framework.sql2o.Dao
 import com.github.vok.framework.sql2o.findAll
 import com.github.vok.framework.sql2o.findById
 import sashastory.dev.model.Application
+import java.math.BigDecimal
 import javax.ws.rs.NotFoundException
 
 /**
@@ -14,6 +15,6 @@ object Sql2oApplicationDao : Dao<Application> {
 
     fun getAllApplications(): List<Application> = findAll()
 
-    fun getApplicationById(id: Long): Application = findById(id) ?: throw NotFoundException("Нет заявки с таким $id")
+    fun getApplicationById(id: BigDecimal): Application = findById(id) ?: throw NotFoundException("Нет заявки с таким $id")
 
 }
